requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai
python-dotenv
tavily-python
logfire
